chrome script/requirements.txt
attrs==23.1.0
backports-datetime-fromisoformat==2.0.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
courlan==0.9.3
cssselect==1.2.0
dateparser==1.1.8
exceptiongroup==1.1.3
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.3
Flask==2.3.3
h11==0.14.0
htmldate==1.5.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.2
jusText==3.0.0
langcodes==3.3.0
langdetect==1.0.9
lxml==4.9.3
MarkupSafe==2.1.3
newspaper3k==0.2.8
nltk==3.8.1
outcome==1.2.0
Pillow==10.0.0
pyahocorasick==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
selenium==4.12.0
selenium-stealth==1.0.6
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tinysegmenter==0.3
tld==0.13
tldextract==3.5.0
tqdm==4.66.1
trafilatura==1.6.1
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==2.0.4
validators==0.21.2
Werkzeug==2.3.7
wsproto==1.2.0
zipp==3.16.2
